# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenPanel API
  description: Self-hosted manga/comic reader server
  version: 0.1.1
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6515
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/auth/login
...
paths:
...
  /api/books/{book_id}/pages/{page_num}:
    get:
      tags: [Reader]
      summary: Serve page image
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
        - name: page_num
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Image data
          content:
            image/*:
              schema:
                type: string
                format: binary
...